➕[Add]: Errors.swift 파일 생성
